Working on proper support for 'this' in the nullability domain.
More or less working, however using some references to access the fields of the current object, which is not ideal.
One potential problem is using a local variable/parameter that shadows a field.  If the variable is not in the environment, it will write to the field instead.
Current mitigation is to put all the local variables in environment as bottom and parameters as top when using the call function.